Added basic IO interaction in the browser
The terminal currently supports an input box that reacts to any commands
written inside of it. When a user types a command, it's parsed and check
if it's valid. Then a new input tag, below the original one, is inserted
and focused. Then the user can continue typing new commands.

Although the previous input tag can be focused, it cannot accpet any new
commands, which behaves exactly like an terminal.

Also added:
* Documentation
* Configurations file
* Basic VFS boilerplate