Added mish-mash CV
This adds the basic functionality and view to the application. The
mish-mash CV application interacts with the user through a CLI and
sends requests to the server in order to parse them. After parsing
them, the VFS does it's job by providing the requested information and
then a response is generated.

Specifically added:
* View component (HTML and CSS)
* Personal information
* VFS functionality
* Configuration file
* Project structure for dependencies